Add an error message for building architecture mismatch
Snow Leopard defaults to 64-bit if you don't specify an architecture,
whereas in the past it defaulted to 32-bit. This isn't a problem for Qt
per-se. It *is* a problem if you just build an application that uses Qt
but isn't using qmake stuff. To help in those situations, we should
error out to let the person know that they need to change their
configuration (in any case, the headers are going to complain and they
get a much more cryptic message).

Reviewed by: Morten Sørvig