BT: QtDemo works far to sluggish on mac
Basically we need to revert more of Oliviers change. The reason
is that we just cannot run the demo on mac without timer based
updates (as opposed to graphics view itself tries to figure out
when to update). The reason is that the demo uses QItemAnimations
extensivelly, which again uses QTimelines, wich again uses
separate timers. And on mac, because of coalesced updated, each
item animation will result in an individual redrawing of the scene.

Task-number: 255751
Reviewed-by: Olivier