qdoc: Added new class names for different tables.
class="valuelist" is for the table used for enum types.
class="alignedsummary" is for summary sections aligned on the name.
class="propsummary" is for the property summary section.
class="toc" is used for tables of contents.
class="generic" is used for all other tables. We might need to break this down more.