Partially revert "Add (back) properties to QGraphicsItem to change the transformations component"
This partially reverts commit 56f23d4c, which changed the logic in
updateCachedClipPathFromSetPosHelper. We cannot compute the item's
transform relative to the parent by using transformToParent() because
the new position is not yet set on the item, that's why the new position
is passed into the function.

However, I'll look into how we can get rid of the entire function, but
keep it as is for now.

Reviewed-by: Olivier