QGraphicsItems are repainted when outside the view's exposed region.
The problem was that we used QRect::isEmpty() on the item's bounding
rect intersected with the exposed region's bounding rect as a criteria
for whether the item should be drawn or not. This does not work as
expected with partial updates, where the boundingRect() of the exposed
region easily can cover the entire viewport area.

The item should *only* be drawn if its bounding rect intersects with
the exposed region (and not the exposed region's bounding rect).

Auto-test included.