Attempt to fix header installation for Phonon.
This is the long-standing issue of whether Phonon headers should be
written with a capital P or a lowercase one. KDE releases of Phonon
had <Phonon/CapitalClassName> whereas Qt 4.4 had <phonon/filename.h>.

I tried to solve this before by adding a Phonon subdir next to phonon
in include/, but that only compounded the error: the presence of two
dirs caused problems and the installation wasn't fixed.

So instead try to place Phonon/ClassName inside include/phonon. And
fix the installation to do it properly: just copy the include/$lib dir
into the target, then overwrite the .h files with the sources from
src/$lib.

Reviewed-by: Marius Storm-Olsen