Scroll the dirty parts of the scroll area in scrollby for Cocoa.
Ah! One less ###!
The scroll functions in Carbon and Cocoa don't scroll the regions that
have been marked dirty. In the past, we killed performance by updating
the whole view. We got a workaround for Carbon in the form of an SPI,
but I wasn't aware of a corresponding item in the NSView API, but it is
there publically and available in 10.5. Fast scrolling in Cocoa now for
people who like to use the keyboard.

Reviewed by: Morten Sørvig