Crash fix for task #255989
Hold a QPointer to the widget with the context menu. After the menu
pops up, the focussed widget looses focus and may be deleted (like
it happens in this case with the item editor). Testing for isNull()
instead of crashing.

It would be great to not even add the "Action" entry to the system menu
but in that moment we don't know that the current focuswidget will
soon disappear. That's still a Todo.