Fix for selection with Shift-Arrow/Shift-Click in QListView's IconMode
This addresses the selection of items using Shift-Arrow or Shift-Click
in QListView's IconMode if the items are in a grid layout. In the case
that the items do not have the same size (e.g., because their text is
wrapped), this commit prevents the unexpected selection of additional
items. New unit tests are included.

Merge-request: 666
Reviewed-by: Olivier Goffart