Improve the speed of QDir, QFileInfo and QDirIterator.
This patch basically avoid to call too often currentFileInfo in
QDirIterator. It replaces the QHash that was overkill in QFileInfo for
caching filenames. The last part is reordering the matchesFilter
to avoid stat as much as possible by using the power of && and filters
that are set on QDirIterator. And it fixes some random "mistake".

I have added a benchmark in QDir which test some use case with
10000 files in a dir.

Written-with: Benjamin
Reviewed-by: phartman