don't try to delete script object snapshots when debugger front-end is deleted
The locals model is technically part of the debugger front-end. When it's
deleted we will have detached (or are in the process of detaching) from
the debugger target anyway, which means that the object snapshots will
be cleaned up by the back-end itself. In any case we shouldn't be sending
debugger commands from a destructor, since we're not in a stable state.
For in-process debugging, we were lucky and this magically worked, but
for out-of-process debugging it caused the debugger to crash.