Refactored gesture api
Rewritten the api almost from scratch, making it simplier and more
flexible at the same time.

The current implementation will not have complex gseturemanager class
inside Qt, but the QGesture base class, which represents both a
gesture recognizer and a gesture itself with a set of properties. A
set of common gestures that can use used in third-party applications
(and in Qt itself internally) is supposed to be found in
qstandardgestures.h, and a base class for user-defined gestures is in
qgesture.h

Gesture implementation for Pan on Windows7 has also been added as a
reference implementation for platform gestures.