Fixes crash of qmake on Windows Vista Business x64 (and undefined behavior on all versions of Windows).
setvbuf handles its arguments differently on Unix and Windows.
Windows uses the size parameter when the given buffer is NULL, which
appears to violate C89/C99.  Giving a size parameter of 0 with a NULL
buffer caused qmake to crash on Windows Vista Business x64.

Windows also can't set line buffering with setvbuf _at all_ according to
the MSDN documentation, so don't bother calling it on Windows.