Prevented QCompleter popup from appearing briefly on Windows.
It turns out that the fix 7bf4512659113f8cc78e72f1c84158ce4f70a526
caused the QCompleter popup from appearing for a split second on
Windows. This is because the popup is shown as a toplevel window
on this platform. On other platforms than Mac and Windows it doesn't
seem to matter, but be on the safe side, the fix is modified to
apply for Mac only.

Reviewed-by: Daniel Molkentin
Task-number: 255374