Update softkeys correctly when no focus widget is present.
In the case where there is no focus widget present, the button group
container should be updated with the softkeys from the currently active
window if it has softkeys. This patch enables this behavior.

Also modify QMainWindow such that the soft keys are updated when the
main window becomes the active window and not only when the menu bar
is set. At the moment, this creates a new QAction and this should be
changed since it will continuously grow the memory.

As a final fix, move the soft key setting out of setFocus() because it
is much to early. At the point, the widget does not know for certain
that it actually will get focus since it could be disabled or have a
focus proxy.