Improved styling of expander arrows with Gtk+
The current implementation was based on the GtkExpander widget.
However we do not have such a widget and the primary use case is for
expanders inside item views. Hence it makes sense to base it on
gtktreeview instead. As I can not reproduce the crash with
any recent versions of cleanice I will remove the workaround as well.

omment and edit as applicable ]---|

Task-number: 256146
Reviewed-by: Thorbjørn