Implement QMenu::separatorsCollapsible on Mac/Cocoa.
Hide duplicate separators and separators at the beginning and end of the
menu. The main challenge here is that separators and "normal" menu items
can be added in any order, for example add all the separators first
and then fill in the items. This means that we have to redo the visibility
decision for separators above and below when adding normal items.

Reviewed-by: Prasanth Ullattil