Fix detection of stlportv5.
Using two concatenated variables as an argument to exists() causes
problems because this can expand to '//' which qmake thinks is a valid
path (if at the beginning), but this is not what we want here. So add
a work around to manually expand the path without the additional '/'
character and then add a comment saying that changing one should change
the other.

Reviewed-by: Miikka Heikkinen