Fix broken auto-test for the index.
Since now items() doesn't use the index then the growingboundingrect
is not updated if you call right after a delete. It's because the timer
is not yet fired (even the processEvent) so you call add, move, remove
which will trigger only one update index so the growingboundingrect will
never change.