Improve softkey implementation and add a default 'Exit' button to RSK.
The previous implementation had some problems where the actual position
of the softkey was determined by it's position in the list. This is
suboptimal since the position should be tied to the role. Actions that
have "negative" conotation should be located on the right soft key
(RSK) while "positive" actions go on the left soft key (LSK). This is
according to the S60 User Interface Guide.

It is also standard practice in S60 to have an 'Exit' button on the RSK
so we add one of those if there is not already a key present there.

Task-number: 256365
Reviewed-by: Markku Luukkainen