Changed enum Qt::Uninitialized to enum Qt::Initialization
enum Qt::Uninitialized contains one value, which is also called
Uninitialized. Because the type and the value used the same name,
wherever the type Qt::Uninitialized was used in a function signature,
you had to include the enum keyword. But qdoc's preprocessor doesn't
like that, so you have to take the enum keyword out of the signatures.
But then the compiler complains because the type and the value look
the same. So I had to change the enum type name to Initialization, so
the compiler can distinguish it from the value name. And qdoc likes
that too.