Fix crash in widget box when using load mode "LoadCustomWidgetsOnly"
When loading the widget box with load mode "LoadCustomWidgetsOnly", it
will create an empty category. This will indirectly cause a resize event
to be posted for the tree widget of the widget box, which in turn will
cause adjustSubListSize() to be called for each category item in the
tree widget. Since the category has not yet been populated, the
assumption in adjustSubListSize() that cat_item->child(0) is non-null
would cause a crash. The fix is to return immediately if the category is
empty.

Reviewed-by: Friedemann Kleint