Multitouch, Cocoa: Added Qt::WA_TouchPadAcceptSingleTouchEvents
By default this is set to false, meaning you will only get multitouch
events. The reason why this is important is that we use the first
touch of a new touch sequence to find out which widget to send
the subsequent touches to. And on a touchpad, you normally want
this to be the widget under the cursor when more than one finger
is pressed on the pad.