Regression against 4.4 in QRectF::operator|.
In 4.4 QRectF handled flat rectangles in the same fashion as QRect does,
but that changed with Lars' and Jo's optmizations done in the falcon
branch. The difference is that the optimized version only checks whether
the width or height is 0, whereas in 4.4 both had to be 0 (isNull()) before
we bailed out.

This regression also introduced a regression in
QGraphicsItem::childrenBoundingRect().

Auto-test included.

Task-number: 254995
Reviewed-by: Lars