Remove these outdated references
This doesn't work, the paths are wrong. It's not properly checking the
arch, and none of our compilers use it. Unsupported ones might, but I'd
rather leave it blank and let them figure out how to do it correctly
then lead them down a blind alley.

Discussed with Morten Sørvig