QGraphicsAnchorLayout: Add convencience "fill" methods
Adding "fillWidth()", "fillHeight()" and "fill()". These convenience
methods simplify the creation of anchor setups where some items
are meant to use the full width and/or height of the layout.

For instance, instead of creating four anchors (bottom, top, left, right)
between an item and the layout, to make it use the full layout area, the
user can call layout->fill(item).

It is also possible to make an item assume the same width or height of
another item, calling layout->fill(firstItem, secondItem).

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>