Implemented parallel simplification, some bugfixes of the previous code.
Currently, the code is not in effect for the simplex solver kicks in
(it crashes), but it is in effect for each layout to check that
simplification and restoring the simplification back again works.
This is currently done in calculateGraphs()

Parts of the code does not read well, especially the detection of the
sequential chunks.