Only instantiate the PICT pasteboard MIME when we have to.
10.6 does a lot of image synthesis for us when we put up a tiff, which
is great because we don't have to do a thing and it will work with all
the applications out there. This means we don't need our PICT convertor
on 10.6 for 32-bit apps. However, this magic doesn't exist in earlier
versions of Mac OS X, so we have to keep it around (along with the
QuickTime symbol resolving) there.

Reviewed-by: Morten Sørvig