make it possible to support queued connections in qtscript
This change doesn't actually add public API for it, but makes it
trivial to add.
On the C++ side, qScriptConnect() would get an overload that
takes an additional argument, the connection type (d'oh, it
should have had a default argument to begin with!).
On the script side, it's a bit more tricky to "overload" the
existing connect(), since it's already "overloaded" (can have
either one or two parameters). Plus, I'd like connect() to be
able to support bind-like functionality so you can pass
additional arguments to it that will be passed to the signal
handler at signal emission time. Oh well, we'll see.