Experimental: allow QSharedPointer to be used with forward declarations that are declared in this file.
The one-definition rule allows the forward declaration appearing below to
apply to code that was earlier. Therefore, if the compiler finds out
how to delete the object, we can allow a QSharedPointer of a forward-
declared-type.

This means the actual problem is just a warning with g++. To catch the
error, we need a separate .cpp file and I'd rather run this as an
external test.