De-export the ThreadEngineBarrier class and de-inline the call to acquire.
(ThreadEngineBarrier was implemented in the header as ThreadEngineSemaphore
in Qt 4.4 and Qt 4.5, and exported for a brief time in the Qt master branch. )

Unfortunately the inline call to acquire means that applications compiled
with Qt 4.4. and Qt 4.5 will contain a copy of the old implementation.
This is BC as long as the memory layout of the ThreadEngineBarrier class
stays the same, which means that we keep the QMutex member variable.

The old version of acquire will no longer be thread-safe since the new
version doesn't lock the mutex, but this is fortunately not a problem;
the only place it's called is in startAsynchronously() before the worker
threads are started.