Fix painting errors in QScrollArea on Mac (Carbon)
This is mainly a stop-gap solution for 4.5.x. It trades painting performance
for correct painting.

Commit 7988d05da changed the opaque test from q->testAttribute(Qt::WA_OpaquePaintEvent)
to qt_widget_private(w)->isOpaque in qt_mac_update_widget_posisiton. This means
we'll do optimized moves in more cases. Unfortunately it also causes painting errors
in some cases (see the task).

Revert the commit for now to put the 4.5 branch in a god shape.

Task-number: 252295
Reviewed-by: nrc