Fixed QPushButton sizeHint recalculation bug.
This patch causes the size hint of QPushButton to be
recalculated in cases where the value of the autoDefault
property may have changed due to changes in the ancestor
chain.
If not explicitly set, the value of the autoDefault property
depends on the presence of a QDialog ancestor.

Note: The new autotest covers two different use cases related to
this behavior.

Reviewed-by: janarve
Task-number: 255581