Always turn on shadows for Cocoa.
Frameless windows wouldn't get shadows in Cocoa, which they do in
Carbon. You can argue over who is more correct, but the fact is they
can't be inconsistent. Since Cocoa is the newcomer, I'm bending that.
Though it would seem useful to have an ability to provide some developer
control over the shadow. At the moment, the only thing we have to ensure
is that we always turn on the shadow.

Task-number: 254725
Reviewed-by: Denis