Avoid revalidating message parameters.
This is a small performance improvement when making a call: we don't
need to validate what we already know to be valid because we either
designed it to be so or because we've already validated.

The D-Bus library unfortunately validates again and there's nothing we
can do about it. But we can avoid doing it twice in our own code.

Reviewed-By: Marius Bugge Monsen