More re-factoring of QGraphicsSceneIndex.
Mostly re-factoring of QGraphicsSceneIndexPrivate::recursive_items_helper
so it can re-use code from the scene (topLevelItemsInStackingOrder). That also
means we'll use the bsp tree in case of NoIndex instead of always looping
through the top-levels. This function is now almost identical to
QGraphicsScenePrivate::drawSubtreeRecursive, so it might be worth
looking into how we can abstract it even more and have one common
recursive function.