Fixed a warning on Symbian.
If the profile has release mode set, QT_NO_DEBUG was defined twice
(once on command line and once in qglobal.h). So now we check for
that. This is better than to ignore debug and release completely,
because there may be other flags that also change when they change.