Use an "int status" extra parameter in property reading/writing.
When calling qt_metacall with the ReadProperty or WriteProperty, the
data is on argv[0] like it was before, but now the QVariant itself is
on argv[1] and there's an extra parameter in argv[2] which the meta
code can use to indicate result.

This allows QtDBus to process properties much more easily. In the case
of property reading, we need to be able to modify the variant itself,
because copying types when we don't have the data isn't very easy.

As for setting, we need to be able to tell setProperty to return true
or false depending on whether we succeeded in setting the property or
not.

Reviewed-By: Kent Hansen
Reviewed-By: Marius Bugge Monsen