Handle application paths larger than MAX_PATH, and fix potential buffer overflow
These days we can easily get very long paths, so we should support
application paths as long as needed.

There was also a potention exploit in that if the path was MAX_PATH or
larger, the string would not be \0 terminated (see MSDN docs for
GetModuleFileName), and thus cause problems in QString::fromWCharArray().

Merge-request: 604
Reviewed-by: Marius Storm-Olsen <marius@trolltech.com>