Improve the code a bit more by using the variant that QMetaProperty creates.
This works for the case of complex types that have to be demarshalled.
We don't need to instantiate a new type because QMetaProperty has already
done that for us.

Also, fix the handling of properties of type variant. I have verified
as well that the sending of those properties on the wire use a
double-variant encoding (i.e., a variant containing a variant
containing some data, the same that Qt 4.5 uses). It's a bit pedantic
and it's hard to use when reading stuff, because you get a QVariant
containing a QDBusVariant which contains data, but I can't change this
anymore.

Reviewed-By: Marius Bugge Monsen