Drag and drop events are not delivered correctly in Cocoa
Drag and drop events should consider the WA_TransparentForMouseEvents
attribute like the mouse events. If this attribute is set for a widget,
the event has to be passed to right widget under mouse. The widget is
identified by calling hitTest. In such cases the leave event has to be
delivered to the widget which actually accepted the enter event.

Task-number: 252088
Reviewed-by: Norwegian Rock Cat