Disable private unit tests when Qt is configured without -developer-build, part 2.
Some autotests use private (unexported) code, either because they're
testing private classes or because that's the easiest way to test
the public classes.  Configuring Qt with `-developer-build' is needed
for these tests.

This commit fixes the tests so configuring without `-developer-build'
only builds the tests which strictly use public API.