Fix to paint big widgets in a scroll area
If a big widget is inside a scroll area, and this widget is in the
limits of XCOORD_MAX, its child might not be inside the limits.
The child is then limited to wrect, but wrect might not be on the
screen because the parent is scrolled. To avoid this problem,
the widgets position should not influence whether wrect is used or not.

Task-number: 144779
Reviewed-by: nrc