Manual cherrypick of QLineControl from kinetic-declarativeui-qfxlineedit
Splits out the control logic of QLineEdit into QLineControl, which is
somewhat similar to QTextEdit. This was originally worked on in the
ianws-qt-lineedit-textedit-research repository, and then further
developed in kinetic in the kinetic-declarativeui-qfxlineedit branch.

Note that this also includes a tiny change to qvalidator.h which allows
validators to be used in qml.