QFontComboBox: used to emit currentFontChanged twice when setting it.
Setting the current font would change the current font and then it
would try to select the right model index and get the font only from
the text of the combobox. This was resetting the point size of the font,
resulting in emitting the signal a second time.
In the case of the user, it was also causing signals to be called in
a loop.

Task-number: 229731