Move the EGL library to a different qmake variable.
The EGL code is now a part of QtGui so QtOpenVG doesn't need to link
against it anymore. This also removes the need to link against cone
since we don't need to call DrawableWindow() from the OpenVG code.
We add the EGL library to a separate variable which is then used when
QT_CONFIG contains the "egl" feature.