Change the inclusion method here to use system includes.
This is a workaround for a bug in the Symbian toolchain where user
includes are included relative to the source file instead of to the
header that actually included it. Making them system includes seems
to work around the issue without introducing problems on other
platforms (so far?). These particular files cause problems because they
are included by modules other than QtGui such as QtOpenVG and QtSVG.