Add OpenVG support to configure.exe.
The configure shell script already had this so add it the .exe so we
can make use of it on other platforms. This adds support for the
'-openvg' option and '-graphicssytem openvg' option and makes sure
the right variables are set in the build system. Building with -openvg
will implicitly enable EGL support as well for now.