Switched to asynchronous focus handling on Symbian.
This was done in order to be more in line with what other platforms
(at least X11) do. In addition, it prevents show() from entering
event handlers in Qt. That should only happen in processEvents().

This required the introduction of a new event,
SymbianDeferredFocusChanged, which we post whenever there is a focus
change.

RevBy:    Jason Barron
AutoTest: Passed