Implement an FSEvents-based QFileSystemWatcherEngine
This has been around for a while and really should have been put in
earlier. Mac OS X (client) has a much lower ulimit for open files than
what some other Unix-like OS's have (in defense it does save memory).
However, if you start watching a lot of files, it will start to fall
down. You can adjust the ulimit, but it's a bit inconvenient. FSEvents
watches the directory and notifies you of changes that happen in that
directory (and below, but we don't really use it). It also can be
adjusted for latency so that performance isn't affected by heavy file
system use (but Qt doesn't use that either at the moment). The other
thing is that it doesn't require any open files, so it's much better for
our number of open files. This feature is only on Leopard and up, so
people wanting to deploy Tiger will still have the "open files" problem
to deal with.

There are still some optimizations available in this code. For example,
we could coalesce things down to watch only one high-level directory
without changing much of the implementation. The current implementation
has some very simplistic ways of handling things, but this simplicity
works well. I documented it, so you can see that, yes, I really
meant to do that.

Task-Id: 164068 (and others)
Reviewed-by: Denis